Move VideoCodecInfo into separate module
